backend/requirements.txt
# FastAPI and Server
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database and Storage
supabase==2.10.0

# LangChain - AI Framework (versiones actualizadas y compatibles)
langchain==0.3.13
langchain-groq==0.2.1
langchain-community==0.3.5
langchain-core==0.3.35
langchain-text-splitters==0.3.6

# Data Validation
pydantic==2.9.2
pydantic-settings==2.6.1

# PDF Processing
pypdf2==3.0.1

# File Upload
python-multipart==0.0.12
aiofiles==24.1.0

# HTTP Client
httpx==0.27.2

# Email Validation
email-validator==2.2.0

# Environment Variables
python-dotenv==1.0.1